Refactor ProgressBar and usePoints hook to dynamically calculate points and update progress; remove unused code in Exchange page
